+++ spec/local/pipelineFeedReaderSpec.js
@@ -48,13 +48,13 @@ describe('pipelineFeedReader Go CD', function () {
 
         thePipelineFeedReader.initFullCache(testPipelines).then(function () {
 
-          thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(results) {
-            expect(_.keys(results).length).toBe(NUM_ENTRIES_IN_FIXTURE);
-            expect(results['2063'].stages.length).toBe(7);
+          thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(results) {
+            expect(_.keys(results.pipelineRuns).length).toBe(NUM_ENTRIES_IN_FIXTURE);
+            expect(results.pipelineRuns['2063'].stages.length).toBe(7);
 
-            thePipelineFeedReader.readPipelineRuns({ pipeline: 'DOWNSTREAM-PIPELINE'}).then(function(downstreamResults) {
-              expect(_.keys(downstreamResults).length).toBe(NUM_ENTRIES_IN_DOWNSTREAM_FIXTURE);
-              expect(downstreamResults['2065'].stages.length).toBe(2);
+            thePipelineFeedReader.readHistory({ pipeline: 'DOWNSTREAM-PIPELINE'}).then(function(downstreamResults) {
+              expect(_.keys(downstreamResults.pipelineRuns).length).toBe(NUM_ENTRIES_IN_DOWNSTREAM_FIXTURE);
+              expect(downstreamResults.pipelineRuns['2065'].stages.length).toBe(2);
               testDone();
             }).done();
 
@@ -67,13 +67,13 @@ describe('pipelineFeedReader Go CD', function () {
 
       it('should cache pipeline run even if its excluded in results when first present', function(testDone) {
         thePipelineFeedReader.initFullCache(testPipelines).then(function () {
-          thePipelineFeedReader.readPipelineRuns({ exclude: ['2066'], pipeline: 'A-PIPELINE' }).then(function(firstResults) {
-            expect(_.keys(firstResults).length).toBe(NUM_ENTRIES_IN_FIXTURE - 1);
+          thePipelineFeedReader.readHistory({ exclude: ['2066'], pipeline: 'A-PIPELINE' }).then(function(firstResults) {
+            expect(_.keys(firstResults.pipelineRuns).length).toBe(NUM_ENTRIES_IN_FIXTURE - 1);
 
             thePipelineFeedReader.initFullCache(testPipelines).then(function () {
-              thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(secondResults) {
-                expect(_.keys(secondResults).length).toBe(NUM_ENTRIES_IN_FIXTURE);
-                expect(secondResults['2066']).toBeDefined();
+              thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(secondResults) {
+                expect(_.keys(secondResults.pipelineRuns).length).toBe(NUM_ENTRIES_IN_FIXTURE);
+                expect(secondResults.pipelineRuns['2066']).toBeDefined();
                 testDone();
               }).done();
 
@@ -86,65 +86,65 @@ describe('pipelineFeedReader Go CD', function () {
 
     });
 
-    describe('readPipelineRuns()', function () {
+    describe('readHistory()', function () {
 
       it('should always read fresh data for at least one page', function (testDone) {
-        thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(results) {
-          expect(results['2066'].stages.length).toBe(7);
-          expect(results['2063']).toBeUndefined();
+        thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(results) {
+          expect(results.pipelineRuns['2066'].stages.length).toBe(7);
+          expect(results.pipelineRuns['2063']).toBeUndefined();
           testDone();
         }).done();
       });
 
       it('should write a sample to a file, for documentation purposes', function (testDone) {
 
-        thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(results) {
+        thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(results) {
           var pipelineRunToLog = '2066';
-          expect(results[pipelineRunToLog]).toBeDefined();
+          expect(results.pipelineRuns[pipelineRunToLog]).toBeDefined();
           var base = path.resolve(__dirname, 'samples');
-          fs.writeFile(base + '/history.json', JSON.stringify(results[pipelineRunToLog], undefined, 2), function() {
+          fs.writeFile(base + '/history.json', JSON.stringify(results.pipelineRuns[pipelineRunToLog], undefined, 2), function() {
             testDone();
           });
         }).done();
       });
 
       it('should initialise a set of pipeline runs', function (testDone) {
-        thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(results) {
-          expect(_.keys(results).length).toBe(NUM_ENTRIES_IN_FIXTURE_PAGE_1);
-          expect(results['2066']).toBeDefined();
+        thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(results) {
+          expect(_.keys(results.pipelineRuns).length).toBe(NUM_ENTRIES_IN_FIXTURE_PAGE_1);
+          expect(results.pipelineRuns['2066']).toBeDefined();
           testDone();
         }).done();
 
       });
 
       it('should exclude specific pipeline runs if specified', function(testDone) {
-        thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE', exclude: ['2066', '2065']}).then(function(results) {
-          expect(_.keys(results).length).toBe(NUM_ENTRIES_IN_FIXTURE_PAGE_1 - 2);
+        thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE', exclude: ['2066', '2065']}).then(function(results) {
+          expect(_.keys(results.pipelineRuns).length).toBe(NUM_ENTRIES_IN_FIXTURE_PAGE_1 - 2);
 
           testDone();
         }).done();
       });
 
       it('should only include specified pipeline names', function(testDone) {
-        thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE-NOT-IN-FIXTURE'}).then(function(results) {
-          expect(_.keys(results).length).toBe(0);
+        thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE-NOT-IN-FIXTURE'}).then(function(results) {
+          expect(_.keys(results.pipelineRuns).length).toBe(0);
 
           testDone();
         }).done();
       });
 
       it('should add stages to respective pipeline runs', function (testDone) {
-        thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(results) {
-          expect(results['2066'].stages.length).toBe(7);
+        thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(results) {
+          expect(results.pipelineRuns['2066'].stages.length).toBe(7);
           testDone();
         }).done();
 
       });
 
       it('should determine the time the last stage got scheduled', function(testDone) {
-        thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(results) {
+        thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(results) {
           var expectedTime = moment(1419000842499);
-          var actualTime = moment(results['2066'].summary.lastScheduled);
+          var actualTime = moment(results.pipelineRuns['2066'].summary.lastScheduled);
           expect(actualTime.hours()).toBe(expectedTime.hours());
           expect(actualTime.minutes()).toBe(expectedTime.minutes());
           expect(actualTime.seconds()).toBe(expectedTime.seconds());
@@ -155,15 +155,15 @@ describe('pipelineFeedReader Go CD', function () {
 
       it('should determine the result of the pipeline', function(testDone) {
         thePipelineFeedReader.initFullCache(testPipelines).then(function () {
-          thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(results) {
-            expect(results['2066'].stages.length).toBe(7);
-            expect(results['2066'].summary.result).toBe('passed');
-            expect(results['2062'].stages.length).toBe(7);
-            expect(results['2062'].summary.result).toBe('failed');
-            expect(results['2062'].summary.stageNotSuccessful).toBe('functional-test');
+          thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(results) {
+            expect(results.pipelineRuns['2066'].stages.length).toBe(7);
+            expect(results.pipelineRuns['2066'].summary.result).toBe('passed');
+            expect(results.pipelineRuns['2062'].stages.length).toBe(7);
+            expect(results.pipelineRuns['2062'].summary.result).toBe('failed');
+            expect(results.pipelineRuns['2062'].summary.stageNotSuccessful).toBe('functional-test');
 
-            thePipelineFeedReader.readPipelineRuns({ pipeline: 'DOWNSTREAM-PIPELINE'}).then(function(resultsDownstream) {
-              expect(resultsDownstream['2066'].summary.result).toBe('cancelled');
+            thePipelineFeedReader.readHistory({ pipeline: 'DOWNSTREAM-PIPELINE'}).then(function(resultsDownstream) {
+              expect(resultsDownstream.pipelineRuns['2066'].summary.result).toBe('cancelled');
 
               testDone();
             }).done();
@@ -173,10 +173,10 @@ describe('pipelineFeedReader Go CD', function () {
 
       it('should determine the author of the latest change that triggered the run', function(testDone) {
         thePipelineFeedReader.initFullCache(testPipelines).then(function () {
-          thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(results) {
-            expect(results['2066'].summary.author).toBeDefined();
-            expect(results['2066'].summary.author.name).toBe('Edward Norton');
-            expect(results['2063'].summary.author.name).toBe('Brad Pitt');
+          thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(results) {
+            expect(results.pipelineRuns['2066'].summary.author).toBeDefined();
+            expect(results.pipelineRuns['2066'].summary.author.name).toBe('Edward Norton');
+            expect(results.pipelineRuns['2063'].summary.author.name).toBe('Brad Pitt');
 
             testDone();
           }).done();
@@ -184,8 +184,8 @@ describe('pipelineFeedReader Go CD', function () {
       });
 
       it('should parse committer and commit message from material HTML, sorted by latest change first', function(testDone) {
-        thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(results) {
-          var changeInfo = results['2066'].summary.changeInfo;
+        thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(results) {
+          var changeInfo = results.pipelineRuns['2066'].summary.changeInfo;
           expect(changeInfo.committer).toContain('Edward Norton');
           expect(changeInfo.comment).toContain('Some comment');
           expect(changeInfo.revision).toBe('cb855ca1516888541722d8c0ed8973792f30ee57');
@@ -195,36 +195,36 @@ describe('pipelineFeedReader Go CD', function () {
       });
 
       it('should put author and commit message of the latest change into info text, if present', function(testDone) {
-        thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(results) {
+        thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(results) {
           var expectedTimeText = moment(1419000842499).format('HH:mm:ss, MMMM Do YYYY');
-          expect(results['2066'].summary.text).toBe('[2066] passed | Edward Norton | Some comment 5554 | ' + expectedTimeText);
+          expect(results.pipelineRuns['2066'].summary.text).toBe('[2066] passed | Edward Norton | Some comment 5554 | ' + expectedTimeText);
 
           testDone();
         }).done();
       });
 
       it('should create initials of person that authored changes for a failed job', function(testDone) {
-        thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(results) {
-          expect(results['2066'].summary.author.initials).toContain('eno');
+        thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(results) {
+          expect(results.pipelineRuns['2066'].summary.author.initials).toContain('eno');
 
           testDone();
         }).done();
       });
 
       it('should read the details about material for each pipeline', function(testDone) {
-        thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(results) {
-          expect(results['2066']['build_cause'].files.length).toBe(17);
+        thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(results) {
+          expect(results.pipelineRuns['2066']['build_cause'].files.length).toBe(17);
 
           testDone();
         }).done();
       });
 
       it("should determine the state of a stage based on its jobs", function(testDone) {
-        thePipelineFeedReader.readPipelineRuns({ pipeline: 'A-PIPELINE'}).then(function(results) {
-          expect(results['2066'].stages[2].summary.state).toBe('Building');
-          expect(results['2065'].stages[2].summary.state).toBe('Scheduled');
-          expect(results['2065'].stages[5].summary.state).toBe('Completed');
-          expect(results['2064'].stages[2].summary.state).toBe('Building');
+        thePipelineFeedReader.readHistory({ pipeline: 'A-PIPELINE'}).then(function(results) {
+          expect(results.pipelineRuns['2066'].stages[2].summary.state).toBe('Building');
+          expect(results.pipelineRuns['2065'].stages[2].summary.state).toBe('Scheduled');
+          expect(results.pipelineRuns['2065'].stages[5].summary.state).toBe('Completed');
+          expect(results.pipelineRuns['2064'].stages[2].summary.state).toBe('Building');
           testDone();
         }).done();
       });
@@ -238,27 +238,27 @@ describe('pipelineFeedReader Go CD', function () {
         });
 
         it("should determine the upstream pipeline run", function(testDone) {
-          thePipelineFeedReader.readPipelineRuns({ pipeline: 'DOWNSTREAM-PIPELINE'}).then(function(results) {
-            expect(results['2066'].summary.upstream.pipelineName).toBe("A-PIPELINE");
-            expect(results['2066'].summary.upstream.pipelineLabel).toBe("2066");
+          thePipelineFeedReader.readHistory({ pipeline: 'DOWNSTREAM-PIPELINE'}).then(function(results) {
+            expect(results.pipelineRuns['2066'].summary.upstream.pipelineName).toBe("A-PIPELINE");
+            expect(results.pipelineRuns['2066'].summary.upstream.pipelineLabel).toBe("2066");
 
             testDone();
           }).done();
         });
 
         it('should determine the author of the latest change that triggered the upstream pipeline', function(testDone) {
-          thePipelineFeedReader.readPipelineRuns({ pipeline: 'DOWNSTREAM-PIPELINE'}).then(function(results) {
-            expect(results['2066'].summary.author).toBeDefined();
-            expect(results['2066'].summary.author.name).toBe('Edward Norton');
+          thePipelineFeedReader.readHistory({ pipeline: 'DOWNSTREAM-PIPELINE'}).then(function(results) {
+            expect(results.pipelineRuns['2066'].summary.author).toBeDefined();
+            expect(results.pipelineRuns['2066'].summary.author.name).toBe('Edward Norton');
 
             testDone();
           }).done();
         });
 
         it('should parse committer and commit message from material HTML, sorted by latest change first', function(testDone) {
-          thePipelineFeedReader.readPipelineRuns({ pipeline: 'DOWNSTREAM-PIPELINE'}).then(function(results) {
+          thePipelineFeedReader.readHistory({ pipeline: 'DOWNSTREAM-PIPELINE'}).then(function(results) {
 
-            var changeInfo = results['2066'].summary.changeInfo;
+            var changeInfo = results.pipelineRuns['2066'].summary.changeInfo;
             expect(changeInfo.committer).toContain('Edward Norton');
             expect(changeInfo.comment).toContain('Some comment');
             expect(changeInfo.revision).toBe('cb855ca1516888541722d8c0ed8973792f30ee57');
@@ -268,13 +268,13 @@ describe('pipelineFeedReader Go CD', function () {
         });
 
         it('should put author and commit message of the latest change into info text, if present', function(testDone) {
-          thePipelineFeedReader.readPipelineRuns({ pipeline: 'DOWNSTREAM-PIPELINE'}).then(function(results) {
+          thePipelineFeedReader.readHistory({ pipeline: 'DOWNSTREAM-PIPELINE'}).then(function(results) {
 
             var expectedTime2065 = moment(1450880504418).format('HH:mm:ss, MMMM Do YYYY');
-            expect(results['2065'].summary.text).toBe('[2065] passed | Edward Norton | Some comment | ' + expectedTime2065);
+            expect(results.pipelineRuns['2065'].summary.text).toBe('[2065] passed | Edward Norton | Some comment | ' + expectedTime2065);
 
             var expectedTime2066 = moment(1450893820238).format('HH:mm:ss, MMMM Do YYYY');
-            expect(results['2066'].summary.text).toBe('[2066] Stage cancelled: smoke-test | Edward Norton | Some comment 5554 | ' + expectedTime2066);
+            expect(results.pipelineRuns['2066'].summary.text).toBe('[2066] Stage cancelled: smoke-test | Edward Norton | Some comment 5554 | ' + expectedTime2066);
 
             testDone();
           }).done();